views.py  classview형식으로 일부 변경 및 오류수정
오류 수정 완료) 스크랩 url, 댓글 띄우기,

+++ likecompetition/posts/views.py
@@ -1,21 +1,48 @@
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.generic.base import View
+from django.views.generic.edit import FormMixin
 from django.views.generic import ListView, DetailView, UpdateView, DeleteView, CreateView, FormView, TemplateView
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models import F
-from django.http import HttpResponse
-from .models import Post, Comment, Scrap, Area #, Tag
+from django.http import HttpResponse, HttpResponseRedirect
+from .models import Post, Comment, Scrap, Area
 from .forms import PostForm, CommentForm
 from dal import autocomplete
 from taggit.models import Tag
 from django.urls import reverse_lazy, reverse
 from likecompetition.views import OwnerOnlyMixin
 from hitcount.views import HitCountDetailView
+from django.contrib import messages
 
-class PostDetailView(HitCountDetailView):
+class PostDetailView(LoginRequiredMixin, FormMixin, DetailView):
     template_name = 'posts/post_detail.html'
     model = Post
-    count_hit = True
+    form_class = CommentForm
+
+    def get_success_url(self):
+        return reverse_lazy('posts:post_detail', kwargs={'pk': self.object.pk})
+
+    def get_context_data(self, **kwargs):	
+        context = super().get_context_data(**kwargs)
+        context['user'] = self.request.user	
+        context['comments'] = self.object.comment_set.all()
+        return context
+
+    def post(self, request, *args, **kwargs):	
+        self.object = self.get_object()	
+        form = self.get_form()		
+
+        if form.is_valid():			
+            return self.form_valid(form)
+        else:			
+            return self.form_invalid(form)
+
+    def form_valid(self, form):
+        comment = form.save(commit=False)	
+        comment.post = get_object_or_404(Post, pk=self.object.pk) 
+        comment.user = self.request.user
+        comment.save()
+        return super(PostDetailView, self).form_valid(form)
 
 class TagAutocompleteView(autocomplete.Select2QuerySetView):
     def get_queryset(self):
@@ -35,6 +62,7 @@ class PostCreateView(LoginRequiredMixin, CreateView):
         return super().form_valid(form)
 
 class PostEditView(OwnerOnlyMixin, UpdateView):
+    template_name = 'posts/post_form.html'
     model = Post
     form_class = PostForm
 
@@ -46,42 +74,38 @@ class PostDeleteView(OwnerOnlyMixin, DeleteView):
     model = Post
     success_url = reverse_lazy('index')
 
-class CommentView(LoginRequiredMixin, View):
-    def post(self, request, *args, **kwargs):
-        post = get_object_or_404(Post, pk=self.kwargs['post_id'])
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.post = post
-            comment.user = request.user
-            comment.save()
-        return redirect('post_detail', post_id=post.id)
+class CommentDeleteView(OwnerOnlyMixin, View):
+    model = Comment
 
-    def delete(self, request, *args, **kwargs):
+    def dispatch(self, request, *args, **kwargs):
         comment = get_object_or_404(Comment, pk=self.kwargs['comment_id'])
         if request.user == comment.user:
             comment.delete()
-        return redirect('post_detail', post_id=comment.post.id)
-
+        return redirect('posts:post_detail', pk=comment.post.pk)
+        
 class ScrapView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         scraps = Scrap.objects.filter(user=request.user).order_by('-date')
         return render(request, 'scrap.html', {'scraps': scraps})
 
     def post(self, request, *args, **kwargs):
-        post_id = self.kwargs['post_id']
-        post = get_object_or_404(Post, pk=post_id)
+        post = get_object_or_404(Post, pk=self.kwargs['post_id'])
         if not Scrap.objects.filter(user=request.user, post=post).exists():
             scrap = Scrap(user=request.user, post=post)
             scrap.save()
-        return HttpResponse()
+        # return redirect('posts:scrap_list')
+        return redirect('posts:post_detail', pk=post.pk)
 
-    def delete(self, request, *args, **kwargs):
+class ScrapDeleteView(OwnerOnlyMixin, DeleteView):
+    model = Scrap
+
+    def dispatch(self, request, *args, **kwargs):
         post = get_object_or_404(Post, pk=self.kwargs['post_id'])
         scrap = get_object_or_404(Scrap, user=request.user, post=post)
-        scrap.delete()
-        return HttpResponse()
-
+        if request.user == scrap.user:
+            scrap.delete()
+        return redirect('posts:scrap_list')
+        
 class LoadAreasView(View):
     def dispatch(self, request, *args, **kwargs):
         if request.method == 'GET':
